Add high EMF (PWM) mode setting for Pixel 10 series
In SettingsGoogle and PixelDisplayService, the array resource lookup for
config_availableEMValueOptions is done dynamically. However, this
resource actually exists in AOSP, so we can just refer to it statically
like all other resources.

The two strings for this option are pulled in via adevtool, but the icon
is added directly in this commit because it contains references to other
resources. The icon is an exact recreation of the stock OS icon. It
turns out Google just used the `visibility` icon from the Apache 2.0
licensed Material Symbols project verbatim.

https://github.com/google/material-design-icons/blob/bb04090f930e272697f2a1f0d7b352d92dfeee43/symbols/android/visibility/materialsymbolsoutlined/visibility_24px.xml

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>